Use python-gpg instead of python-gpgme if it's installed
I think we need to keep gpgme support because we only build a single .deb
for all the distros we ship to.

I removed python-gpgme from the "Suggests" line in the deb and replaced it
with python-gpg because I don't think there's harm in suggesting a package
that doesn't exist for Ubuntu 17.04 and lower.

I didn't touch the rpm because I couldn't figure out if python-gpg exists
on Fedora. It probably does, but my google-foo is failing me.

Fixes #51