Revert "fix #15" (launching dropbox with dbus-launch)
Launching dropbox with `dbus-launch` makes it unable to communicate
with the shell for the tray icon that uses Unity.

This fixes #53, but it undoes the fix for #15. Since #53 affects
anyone using the default shell that Ubuntu ships with, we should
revert the diff to fix them and then figure out a more robust way to
fix #15.

This reverts commit 0d81fefb08dabe3470964293f7bf3b1de79c4b3c.