build: handle (in some way) the new GNOME versioning scheme
With GNOME switching from odd/even versioning for unsstable/stable
branches to 40.alpha, 40.beta, we need to trick the versioning to
our liking.

In case NAUTILUS_VERSION_MINOR contains letters, we reset it to 0.

This means we treat 40.alpha as 40.0 (as if it was the final release of the 40 series)

Fixes issue#97