Update the build and packaging script for Fedora 37. (#109)
Fedora 37 includes GNOME 43, which is not backwards compatible with packages built against older versions. We used to build new Fedora versions on Ubuntu, but they removed support for building RPM packages in 20.04. By updating the build script to run on Fedora, this allows us to release coreyberla's fix, resolving #108 and #101.

Co-authored-by: Geoffry Song <goffrie@dropbox.com>