Prevent single-item submenus
Nautilus 43+ uses Gtk4 context menus, where submenus have to be
clicked on and entered instead of opening on hover. This makes
accessing submenus containing only a single item needlessly tedious.

So, we detect when the Dropbox submenu will only contain a single
selection, and promote it up to the top level instead of burying
it within an unnecessary submenu layer.